Add a somewhat lengthy C++ implementation
An overly complicated implementation of FizzBuzz in C++. Some stuff
that can be found in this implementation: duplicated code, unused
variables, unused imports, a severe case of NIH syndrome, and
apparently a lack of understanding of some of C++’s features (and
common sense).

Still, it gets the job done.