Add a somewhat wheel-reinventing Ruby implementation
An eyebrow-raising implementation of FizzBuzz in Ruby.  Constructs a
list of (more than) 100 numbers by first concatenating number strings
to numbers, and then converting the first 100 natural numbers of that
list into a list of integers, which is then finally used for the
FizzBuzz magic.

It gets the job done reasonably well.