Add a somewhat double-spaced Lua implementation
A quick and dirty implementation of FizzBuzz in Lua. Lua normally is
very simple to read and use, but in this particular example a
ridiculous number of newlines have been added, making it quite hard to
read (which according to many non-programmers is pretty much the
definition of “code” anyway).